fix: dependency updates (#60)
Updates the majority of the dependencies. These are the commits in
order:

- chore: npm install --save-dev serve@14.2.4
- chore: npm install --save-dev nodemon@3.1.10
- chore: npm install --save-dev autoprefixer@10.4.21
- fix: npm install --save-dev postcss@8.5.6 postcss-cli@11.0.1
- chore: npm install @11ty/eleventy@3.1.2
- fix: npm run css
- chore: npm install --save-dev @commitlint/cli@19.8.1
@commitlint/config-conventional@19.8.1
- chore: npm install --save-dev husky@9.1.7
- chore: remove the following two lines from .husky/commit-msg
- chore: husky install -> husky 
- chore: npm install --save-dev stylelint@16.21.0
- fix: npm run css-lint -- --fix

These are all updates to dependencies that can be tested during local
dev, with the exception of gh-pages. We'll find out about that one...

However, if you run this up locally, you can use `diffsite` to check the
visual differences against the published pages, e.g.:


https://pianomister.github.io/diffsite/?url1=http%3A%2F%2Flocalhost%3A8080%2Fbootstrap-theme%2F%23alert&url2=https%3A%2F%2Ftechfromsage.github.io%2Fbootstrap-theme%2F