chore: pin SSH key for CircleCI (#73)
We use a read/write deploy key from CircleCI, to release changes to
GitHub. However, this means we do not have a default read-only deploy
key - and _that_ means that each new person who touches the project will
generate a new deploy key, breaking the build.

Here, we pin the read/write key, so that we can create a read-only
deploy key without breaking the build.